refactor: consolidate progress modules
