+++ src/my_unicorn/core/progress/__init__.py
@@ -4,8 +4,29 @@ This module contains all progress bar, spinner,
 and progress display logic for my-unicorn.
 """
 
-from my_unicorn.core.progress.progress import ProgressDisplay
+from my_unicorn.core.progress.progress import (
+    IDGenerator,
+    LoggerSuppression,
+    ProgressDisplay,
+    SessionManager,
+    TaskInfo,
+    TaskRegistry,
+    TaskState,
+    TaskConfig,
+    create_api_fetching_task,
+    create_installation_workflow,
+    create_verification_task,
+)
 
 __all__ = [
+    "IDGenerator",
+    "LoggerSuppression",
     "ProgressDisplay",
+    "SessionManager",
+    "TaskInfo",
+    "TaskRegistry",
+    "TaskState",
+    "create_api_fetching_task",
+    "create_installation_workflow",
+    "create_verification_task",
 ]
