--- src/my_unicorn/core/progress/ascii_format.py
@@ -1,54 +1,0 @@
-"""Pure formatting helper functions for ASCII progress display.
-
-This module contains pure functions (no state dependencies) extracted from
-AsciiProgressBackend for better testability and reusability.
-"""
-
-from __future__ import annotations
-
-import time
-
-from .progress_types import SPINNER_FRAMES, TaskState
-
-
-def compute_display_name(task: TaskState) -> str:
-    """Return a condensed display name for a task (strip extension).
-
-    Args:
-        task: The task state containing the name.
-
-    Returns:
-        The task name with .AppImage extension stripped if present.
-    """
-    name = task.name
-    if name.endswith(".AppImage"):
-        return name[:-9]
-    return name
-
-
-def compute_spinner(fps: int) -> str:
-    """Compute the current spinner frame based on time and FPS.
-
-    Args:
-        fps: Frames per second for spinner animation.
-
-    Returns:
-        The current spinner frame character.
-    """
-    current_time = time.monotonic()
-    spinner_idx = int(current_time * fps) % len(SPINNER_FRAMES)
-    return SPINNER_FRAMES[spinner_idx]
-
-
-def compute_download_header(download_count: int) -> str:
-    """Return the downloads section header string.
-
-    Args:
-        download_count: The total number of downloads.
-
-    Returns:
-        A formatted header string for the downloads section.
-    """
-    if download_count > 1:
-        return f"Downloading ({download_count}):"
-    return "Downloading:"
