+++ tests/core/post_download/test_progress_and_verification.py
@@ -12,7 +12,7 @@ from my_unicorn.core.post_download import (
     PostDownloadContext,
     PostDownloadProcessor,
 )
-from my_unicorn.core.progress.display import ProgressDisplay
+from my_unicorn.core.progress import ProgressDisplay
 from my_unicorn.core.protocols.progress import NullProgressReporter
 from my_unicorn.exceptions import VerificationError
 
@@ -32,7 +32,7 @@ class TestSetupProgressTracking:
         processor_instance.progress_reporter = mock_display
 
         with patch(
-            "my_unicorn.core.progress.display_workflows"
+            "my_unicorn.core.progress"
             ".create_installation_workflow",
             new_callable=AsyncMock,
         ) as mock_create_workflow:
@@ -77,7 +77,7 @@ class TestSetupProgressTracking:
         processor_instance.progress_reporter = mock_display
 
         with patch(
-            "my_unicorn.core.progress.display_workflows"
+            "my_unicorn.core.progress"
             ".create_installation_workflow",
             new_callable=AsyncMock,
         ) as mock_create_workflow:
