+++ tests/core/progress/test_display_session.py
@@ -9,8 +9,8 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from my_unicorn.core.progress.display_id import IDGenerator
-from my_unicorn.core.progress.display_session import SessionManager
+from my_unicorn.core.progress import IDGenerator
+from my_unicorn.core.progress import SessionManager
 from my_unicorn.core.progress.progress_types import (
     ProgressConfig,
     ProgressType,
